Revert "Fixed and renamed services navigators. Added logic for handling not found paths"
