+++ src/App.js
@@ -2,7 +2,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom'
 import AppNavigator from '~/navigators/AppNavigator'
 import AuthNavigator from '~/navigators/AuthNavigator'
 import { MantineProvider } from '@mantine/core'
-import ServicesNavigator from '~/navigators/ServicesNavigator'
+import NotFoundRoutes from '~/navigators/notFound'
 
 const App = () => {
   return (
@@ -11,7 +11,7 @@ const App = () => {
         <Routes>
           <Route path="/*" element={<AppNavigator />} />
           <Route path="auth/*" element={<AuthNavigator />} />
-          <Route path="services/*" element={<ServicesNavigator />} />
+          <Route path="services/*" element={<NotFoundRoutes />} />
         </Routes>
       </MantineProvider>
     </BrowserRouter>
