+++ src/__constants__/paths/index.js
@@ -1,3 +1,2 @@
 export { default as APP_PATHS } from './appPaths'
 export { default as AUTH_PATHS } from './authPaths'
-export { default as SERVICES_PATHS } from './servicesPaths'
