--- src/__constants__/paths/servicesPaths.js
@@ -1,5 +1,0 @@
-const SERVICES_PATHS = {
-  NOT_FOUND: '/404'
-}
-
-export default SERVICES_PATHS
