--- src/navigators/ServicesNavigator/index.js
@@ -1,3 +1,0 @@
-import ServiceRoutes from './ServicesNavigator'
-
-export default ServiceRoutes
