+++ src/navigators/notFound/index.js
@@ -1,0 +1,3 @@
+import ServiceRoutes from './notFoundNavigator'
+
+export default ServiceRoutes
