--- src/pages/Services/NotFound/index.js
@@ -1,1 +1,0 @@
-export { default as NotFound } from './NotFound'
