--- src/pages/Services/index.js
@@ -1,1 +1,0 @@
-export * from './NotFound'
