+++ src/pages/Services/notFound/notFound.js
@@ -1,0 +1,7 @@
+import NotFoundForm from './notFoundForm'
+
+const NotFound = () => {
+  return <NotFoundForm />
+}
+
+export default NotFound
